requirements.txt
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
openai==1.13.3
python-dotenv==1.0.1
streamlit==1.31.1